Add nonces to import/export functions

+++ disqus/disqus.php
@@ -416,55 +416,67 @@ function dsq_request_handler() {
                 if (current_user_can('manage_options') && DISQUS_CAN_EXPORT) {
                     $timestamp = intval($_GET['timestamp']);
                     $post_id = intval($_GET['post_id']);
-                    global $wpdb, $dsq_api;
-                    $post = $wpdb->get_results($wpdb->prepare("
-                        SELECT *
-                        FROM $wpdb->posts
-                        WHERE post_type != 'revision'
-                        AND post_status = 'publish'
-                        AND comment_count > 0
-                        AND ID > %d
-                        ORDER BY ID ASC
-                        LIMIT 1
-                    ", $post_id));
-                    $post = $post[0];
-                    $post_id = $post->ID;
-                    $max_post_id = $wpdb->get_var($wpdb->prepare("
-                        SELECT MAX(ID)
-                        FROM $wpdb->posts
-                        WHERE post_type != 'revision'
-                        AND post_status = 'publish'
-                        AND comment_count > 0
-                    ", $post_id));
-                    $eof = (int)($post_id == $max_post_id);
-                    if ($eof) {
-                        $status = 'complete';
-                        $msg = dsq_i('Your comments have been sent to Disqus and queued for import!<br/><a href="'.DISQUS_IMPORTER_URL.'" target="_blank">See the status of your import at Disqus</a>');
+                    if ( isset($_GET['_dsqexport_wpnonce']) === false ) {
+                        $msg = dsq_i('Unable to export comments. Make sure you are accessing this page from the Wordpress dashboard.');
+                        $result = 'fail';
+                        $response = null;
                     }
                     else {
-                        $status = 'partial';
-                        $msg = dsq_i('Processed comments on post #%s&hellip;', $post_id);
-                    }
-                    $result = 'fail';
-                    $response = null;
-                    if ($post) {
-                        require_once(dirname(__FILE__) . '/export.php');
-                        $wxr = dsq_export_wp($post);
-                        $response = $dsq_api->import_wordpress_comments($wxr, $timestamp, $eof);
-                        if (!($response['group_id'] > 0)) {
-                            $result = 'fail';
-                            $msg = '<p class="status dsq-export-fail">'. dsq_i('Sorry, something unexpected happened with the export. Please <a href="#" id="dsq_export_retry">try again</a></p><p>If your API key has changed, you may need to reinstall Disqus (deactivate the plugin and then reactivate it). If you are still having issues, refer to the <a href="%s" onclick="window.open(this.href); return false">WordPress help page</a>.', 'http://disqus.com/help/wordpress'). '</p>';
-                            $response = $dsq_api->get_last_error();
+
+                        // Check nonce
+                        check_admin_referer('dsq-wpnonce_export', '_dsqexport_wpnonce');
+
+                        global $wpdb, $dsq_api;
+                        $post = $wpdb->get_results($wpdb->prepare("
+                            SELECT *
+                            FROM $wpdb->posts
+                            WHERE post_type != 'revision'
+                            AND post_status = 'publish'
+                            AND comment_count > 0
+                            AND ID > %d
+                            ORDER BY ID ASC
+                            LIMIT 1
+                        ", $post_id));
+                        $post = $post[0];
+                        $post_id = $post->ID;
+                        $max_post_id = $wpdb->get_var($wpdb->prepare("
+                            SELECT MAX(ID)
+                            FROM $wpdb->posts
+                            WHERE post_type != 'revision'
+                            AND post_status = 'publish'
+                            AND comment_count > 0
+                        ", $post_id));
+                        $eof = (int)($post_id == $max_post_id);
+                        if ($eof) {
+                            $status = 'complete';
+                            $msg = dsq_i('Your comments have been sent to Disqus and queued for import!<br/><a href="'.DISQUS_IMPORTER_URL.'" target="_blank">See the status of your import at Disqus</a>');
                         }
                         else {
-                            if ($eof) {
-                                $msg = dsq_i('Your comments have been sent to Disqus and queued for import!<br/><a href="%s" target="_blank">See the status of your import at Disqus</a>', $response['link']);
-
+                            $status = 'partial';
+                            $msg = dsq_i('Processed comments on post #%s&hellip;', $post_id);
+                        }
+                        $result = 'fail';
+                        $response = null;
+                        if ($post) {
+                            require_once(dirname(__FILE__) . '/export.php');
+                            $wxr = dsq_export_wp($post);
+                            $response = $dsq_api->import_wordpress_comments($wxr, $timestamp, $eof);
+                            if (!($response['group_id'] > 0)) {
+                                $result = 'fail';
+                                $msg = '<p class="status dsq-export-fail">'. dsq_i('Sorry, something unexpected happened with the export. Please try again.</p><p>If your API key has changed, you may need to reinstall Disqus (deactivate the plugin and then reactivate it). If you are still having issues, refer to the <a href="%s" onclick="window.open(this.href); return false">WordPress help page</a>.', 'http://disqus.com/help/wordpress'). '</p>';
+                                $response = $dsq_api->get_last_error();
                             }
-                            $result = 'success';
+                            else {
+                                if ($eof) {
+                                    $msg = dsq_i('Your comments have been sent to Disqus and queued for import!<br/><a href="%s" target="_blank">See the status of your import at Disqus</a>', $response['link']);
+
+                                }
+                                $result = 'success';
+                            }
                         }
                     }
-// send AJAX response
+                    
+                    // send AJAX response
                     $response = compact('result', 'timestamp', 'status', 'post_id', 'msg', 'eof', 'response');
                     header('Content-type: text/javascript');
                     echo cf_json_encode($response);
@@ -473,38 +485,49 @@ function dsq_request_handler() {
             break;
             case 'import_comments':
                 if (current_user_can('manage_options')) {
-                    if (!isset($_GET['last_comment_id'])) $last_comment_id = false;
-                    else $last_comment_id = $_GET['last_comment_id'];
-
-                    if ($_GET['wipe'] == '1') {
-                        $wpdb->query("DELETE FROM `".$wpdb->prefix."commentmeta` WHERE meta_key IN ('dsq_post_id', 'dsq_parent_post_id')");
-                        $wpdb->query("DELETE FROM `".$wpdb->prefix."comments` WHERE comment_agent LIKE 'Disqus/%%'");
-                    }
-
-                    ob_start();
-                    $response = dsq_sync_forum($last_comment_id, true);
-                    $debug = ob_get_clean();
-                    if (!$response) {
-                        $status = 'error';
+                    if ( isset($_GET['_dsqimport_wpnonce']) === false ) {
+                        $msg = dsq_i('Unable to import comments. Make sure you are accessing this page from the Wordpress dashboard.');
                         $result = 'fail';
-                        $error = $dsq_api->get_last_error();
-                        $msg = '<p class="status dsq-export-fail">'.dsq_i('There was an error downloading your comments from Disqus.').'<br/>'.htmlspecialchars($error).'</p>';
-                    } else {
-                        list($comments, $last_comment_id) = $response;
-                        if (!$comments) {
-                            $status = 'complete';
-                            $msg = dsq_i('Your comments have been downloaded from Disqus and saved in your local database.');
-                        } else {
-                            $status = 'partial';
-                            $msg = dsq_i('Import in progress (last post id: %s) &hellip;', $last_comment_id);
-                        }
-                        $result = 'success';
+                        $response = null;
                     }
-                    $debug = explode("\n", $debug);
-                    $response = compact('result', 'status', 'comments', 'msg', 'last_comment_id', 'debug');
-                    header('Content-type: text/javascript');
-                    echo cf_json_encode($response);
-                    die();
+                    else
+                    {
+                        // Check nonce
+                        check_admin_referer('dsq-wpnonce_import', '_dsqimport_wpnonce');
+
+                        if (!isset($_GET['last_comment_id'])) $last_comment_id = false;
+                        else $last_comment_id = $_GET['last_comment_id'];
+
+                        if ($_GET['wipe'] == '1') {
+                            $wpdb->query("DELETE FROM `".$wpdb->prefix."commentmeta` WHERE meta_key IN ('dsq_post_id', 'dsq_parent_post_id')");
+                            $wpdb->query("DELETE FROM `".$wpdb->prefix."comments` WHERE comment_agent LIKE 'Disqus/%%'");
+                        }
+
+                        ob_start();
+                        $response = dsq_sync_forum($last_comment_id, true);
+                        $debug = ob_get_clean();
+                        if (!$response) {
+                            $status = 'error';
+                            $result = 'fail';
+                            $error = $dsq_api->get_last_error();
+                            $msg = '<p class="status dsq-export-fail">'.dsq_i('There was an error downloading your comments from Disqus.').'<br/>'.htmlspecialchars($error).'</p>';
+                        } else {
+                            list($comments, $last_comment_id) = $response;
+                            if (!$comments) {
+                                $status = 'complete';
+                                $msg = dsq_i('Your comments have been downloaded from Disqus and saved in your local database.');
+                            } else {
+                                $status = 'partial';
+                                $msg = dsq_i('Import in progress (last post id: %s) &hellip;', $last_comment_id);
+                            }
+                            $result = 'success';
+                        }
+                        $debug = explode("\n", $debug);
+                        $response = compact('result', 'status', 'comments', 'msg', 'last_comment_id', 'debug');
+                        header('Content-type: text/javascript');
+                        echo cf_json_encode($response);
+                        die();
+                    }
                 }
             break;
         }
@@ -1006,91 +1029,101 @@ jQuery(function($) {
     dsq_fire_export();
     dsq_fire_import();
 });
-dsq_fire_export = function() {
-    var $ = jQuery;
-    $('#dsq_export a.button, #dsq_export_retry').unbind().click(function() {
-        $('#dsq_export').html('<p class="status"></p>');
-        $('#dsq_export .status').removeClass('dsq-export-fail').addClass('dsq-exporting').html('Processing...');
-        dsq_export_comments();
-        return false;
+
+var dsq_fire_export = function() {
+    jQuery(function($) {
+        $('#dsq_export a.button').unbind().click(function() {
+            $('#dsq_export .status').removeClass('dsq-export-fail').addClass('dsq-exporting').html('Processing...');
+            dsq_export_comments();
+            return false;
+        });
     });
-}
-dsq_export_comments = function() {
-    var $ = jQuery;
-    var status = $('#dsq_export .status');
-    var export_info = (status.attr('rel') || '0|' + (new Date().getTime()/1000)).split('|');
-    $.get(
-        '<?php echo admin_url('index.php'); ?>',
-        {
-            cf_action: 'export_comments',
-            post_id: export_info[0],
-            timestamp: export_info[1]
-        },
-        function(response) {
-            switch (response.result) {
-                case 'success':
-                    status.html(response.msg).attr('rel', response.post_id + '|' + response.timestamp);
-                    switch (response.status) {
-                        case 'partial':
-                            dsq_export_comments();
+};
+
+var dsq_export_comments = function() {
+    jQuery(function($) {
+        var status = $('#dsq_export .status');
+        var nonce = $('#dsq-form_nonce_export').val();
+        var export_info = (status.attr('rel') || '0|' + (new Date().getTime()/1000)).split('|');        
+        $.get(
+            '<?php echo admin_url("index.php"); ?>',
+            {
+                cf_action: 'export_comments',
+                post_id: export_info[0],
+                timestamp: export_info[1],
+                _dsqexport_wpnonce: nonce
+            },
+            function(response) {
+                switch (response.result) {
+                    case 'success':
+                        status.html(response.msg).attr('rel', response.post_id + '|' + response.timestamp);
+                        switch (response.status) {
+                            case 'partial':
+                                dsq_export_comments();
                             break;
-                        case 'complete':
-                            status.removeClass('dsq-exporting').addClass('dsq-exported');
+                            case 'complete':
+                                status.removeClass('dsq-exporting').addClass('dsq-exported');
                             break;
-                    }
-                break;
-                case 'fail':
-                    status.parent().html(response.msg);
-                    dsq_fire_export();
-                break;
-            }
-        },
-        'json'
-    );
-}
-dsq_fire_import = function() {
+                        }
+                    break;
+                    case 'fail':
+                        status.parent().html(response.msg);
+                        dsq_fire_export();
+                    break;
+                }
+            },
+            'json'
+        );
+    });
+};
+
+var dsq_fire_import = function() {
     var $ = jQuery;
     $('#dsq_import a.button, #dsq_import_retry').unbind().click(function() {
         var wipe = $('#dsq_import_wipe').is(':checked');
-        $('#dsq_import').html('<p class="status"></p>');
         $('#dsq_import .status').removeClass('dsq-import-fail').addClass('dsq-importing').html('Processing...');
         dsq_import_comments(wipe);
         return false;
     });
-}
-dsq_import_comments = function(wipe) {
-    var $ = jQuery;
-    var status = $('#dsq_import .status');
-    var last_comment_id = status.attr('rel') || '0';
-    $.get(
-        '<?php echo admin_url('index.php'); ?>',
-        {
-            cf_action: 'import_comments',
-            last_comment_id: last_comment_id,
-            wipe: (wipe ? 1 : 0)
-        },
-        function(response) {
-            switch (response.result) {
-                case 'success':
-                    status.html(response.msg).attr('rel', response.last_comment_id);
-                    switch (response.status) {
-                        case 'partial':
-                            dsq_import_comments(false);
-                            break;
-                        case 'complete':
-                            status.removeClass('dsq-importing').addClass('dsq-imported');
-                            break;
-                    }
-                break;
-                case 'fail':
-                    status.parent().html(response.msg);
-                    dsq_fire_import();
-                break;
-            }
-        },
-        'json'
-    );
-}
+};
+
+var dsq_import_comments = function(wipe) {
+    jQuery(function($) {
+        var status = $('#dsq_import .status');
+        var nonce = $('#dsq-form_nonce_import').val();
+        var last_comment_id = status.attr('rel') || '0';
+        $.get(
+            '<?php echo admin_url('index.php'); ?>',
+            {
+                cf_action: 'import_comments',
+                last_comment_id: last_comment_id,
+                wipe: (wipe ? 1 : 0),
+                _dsqimport_wpnonce: nonce
+            },
+            function(response) {
+                switch (response.result) {
+                    case 'success':
+                        status.html(response.msg).attr('rel', response.last_comment_id);
+                        switch (response.status) {
+                            case 'partial':
+                                dsq_import_comments(false);
+                                break;
+                            case 'complete':
+                                status.removeClass('dsq-importing').addClass('dsq-imported');
+                                break;
+                        }
+                    break;
+                    case 'fail':
+                        status.parent().html(response.msg);
+                        dsq_fire_import();
+                    break;
+                }
+            },
+            'json'
+        );
+    });
+};
+
 </script>
 <?php
 // HACK: Our own styles for older versions of WordPress.
